requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-Cors==6.0.1
Flask-Migrate==4.0.4
gunicorn==21.2.0
python-dotenv==1.2.1
SQLAlchemy==2.0.45
psycopg2-binary
requests==2.32.5
ummalqura
groq
python-docx
alembic==1.11.1